+++ tests/testthat/test-identifiers.R
@@ -22,5 +22,6 @@ test_that("GUIDs and Waterbutler IDs are detected in OSF URLs", {
 })
 
 test_that("special identifier 'me' is recognized", {
+  skip_if_no_pat()
   expect_match(id_type("me"), "users")
 })