+++ tests/testthat/test-osf_tbl.R
@@ -15,14 +15,17 @@ test_that("empty list returns empty osf_tbl", {
 
 context("osf_tbl validation")
 
-user_tbl <- osf_retrieve_user("me")
-user_tbl$foo <- "bar"
-
 test_that("valid osf_tbls are passed through validation", {
+  skip_on_production_server()
+
+  user_tbl <<- osf_retrieve_user("dguxh")
+  user_tbl$foo <<- "bar"
   expect_s3_class(rebuild_osf_tbl(user_tbl), "osf_tbl_user")
 })
 
 test_that("osf_tbls missing required columns are detected", {
+  skip_on_production_server()
+
   expect_true(has_osf_tbl_colnames(user_tbl))
   expect_false(has_osf_tbl_colnames(user_tbl[-1]))
   expect_false(has_osf_tbl_colnames(user_tbl[-2]))
@@ -30,14 +33,17 @@ test_that("osf_tbls missing required columns are detected", {
 })
 
 test_that("osf_tbls with incorrect column types are detected", {
+  skip_on_production_server()
+
   user_tbl$id <- as.factor(user_tbl$id)
   expect_false(has_osf_tbl_coltypes(user_tbl))
 })
 
-proj_tbl <- osf_retrieve_node("brfza")
-
 test_that("can't combine osf_tbls with different subclasses", {
-  proj_tbl$foo <- "barr"
+  skip_on_production_server()
+
+  proj_tbl <- osf_retrieve_node("brfza")
+  proj_tbl$foo <- "bar"
   out <- rbind(user_tbl, proj_tbl)
   expect_identical(class(out), c("tbl_df", "tbl", "data.frame"))
 })