+++ NEWS.md
@@ -1,3 +1,18 @@
+# osfr 0.2.2
+
+## New functions
+
+* `osf_mv()` to move files and directories to a new project, component, or
+subdirectory
+
+## New features
+
+* `osf_rm()` can now delete files and directories
+
+## Minor improvements and fixes
+
+* Restructured tests to better handle environments in which `OSF_PAT` and/or `OSF_SERVER` are not defined
+
 # osfr 0.2.1
 
 * Minor tweaks to the website